finished general model tui

+++ main.py
@@ -11,11 +11,23 @@ current_f_index = None
 cache_dir_path = os.path.join(os.getcwd(), "cache")
 cartridge_dir_path = os.path.join(os.getcwd(), "cartridges")
 data_dir_path = os.path.join(os.getcwd(), "data")
-
-# NOTE(liam): path is expected
+chat_use_CAG = False
+chat_prompt = """
+You are an AI Language Tutor.
+Generate a simple test based on the following information.
+For each question generated, include both question AND the answer, and
+clearly label them 'QUESTION' and 'ANSWER'.
+If an answer contains an explanation, make the explanation general, and assume
+the test-taker cannot see the context you are provided with.
+"""
 
 
 def add_command(cartridge_id: str = None, *file_names: List[str]):
+    """
+    Adds one or more valid (pdf/json) files to a cartridge.
+    The cartridge id must be specified, and will create a new
+    directory in the cartridge directory if not found.
+    """
     global last_cartridge_id, current_f_index
 
     cartridge_id = cartridge_id or last_cartridge_id
@@ -55,6 +67,12 @@ def add_command(cartridge_id: str = None, *file_names: List[str]):
 
 
 def compile_command(cartridge_id: str = None):
+    """
+    Prepares a cartridge to be used by the program.
+    The specified cartridge id is expected to already exist
+    within the cartridge directory.
+    Outputs a 'cart' file.
+    """
     global current_f_index
 
     cartridge_id = cartridge_id or last_cartridge_id
@@ -76,6 +94,10 @@ def compile_command(cartridge_id: str = None):
 
 
 def load_command(cartridge_id: str = None, cartridge_path: str = None, output_dir: str = cache_dir_path):
+    """
+    Loads a valid cartridge file into the cache directory.
+    Also automatically sets the cartridge id to be used.
+    """
     global last_cartridge_id
 
     cartridge_id = cartridge_id or last_cartridge_id
@@ -95,6 +117,12 @@ def load_command(cartridge_id: str = None, cartridge_path: str = None, output_di
 
 
 def set_command(cartridge_id: str = None):
+    """
+    Manually set the cartridge id to be used.
+    If no arguments are passed, an interactive process
+    will guide user to choose from the available cartridges
+    already loaded in the cache directory.
+    """
     global last_cartridge_id
 
     possible_ids = []
@@ -119,36 +147,36 @@ def set_command(cartridge_id: str = None):
 
 
 def unload_command(cartridge_id: str = None):
+    """
+    Clears out the cache directory.
+    """
+    print("This will clear out all files within the cache. Continue?")
 
-    shutil.rmtree(os.path.join(cache_dir_path, cartridge_id))
+    answer = input("(y/N): ")
+    if answer.lower() == 'y':
+        shutil.rmtree(os.path.join(cache_dir_path, cartridge_id))
+        print("Unloaded '{}' from cache!".format(cartridge_id))
 
-    print("Unloaded '{}' from cache!".format(cartridge_id))
+def prompt_command(*new_prompt):
+    """
+    Simple setter for the prompt used by the chat model.
+    """
+    global chat_prompt
 
-
-def generate_command(*query):
-    global last_cartridge_id
-
-    cartridge_id = last_cartridge_id
-    if cartridge_id is None:
-        print("ERROR: cartridge id was not previously set via 'load'.")
+    if new_prompt is None:
+        print("WARN: no prompt provided. Nothing changed.")
         return
 
-    cartridge_dir_path = os.path.join(cache_dir_path, cartridge_id)
-
-    query = " ".join(query)
-
-    context_lang = "Language: {}\n".format(cartridge_id)
-    context_info = prompter.perform_rag(query, os.path.join(cartridge_dir_path, "faiss.index"),
-                                        os.path.join(cartridge_dir_path, "chunks.txt"))
-
-    context = context_lang + context_info
-    result = prompter.generate_response(context)
-
-    print("Generated:\n" + result)
+    old_prompt = chat_prompt
+    chat_prompt = " ".join(new_prompt)
+    print("prompt: '{}' -> '{}'.".format(old_prompt, chat_prompt))
 
 
 def context_command(*query):
-    global last_cartridge_id
+    """
+    Generates a context from a loaded cartridge based on a given query.
+    """
+    global last_cartridge_id, chat_use_CAG
 
     cartridge_id = last_cartridge_id
     if cartridge_id is None:
@@ -157,22 +185,67 @@ def context_command(*query):
 
     cartridge_dir_path = os.path.join(cache_dir_path, cartridge_id)
 
-    query = " ".join(query)
+    query_joined = " ".join(query)
 
     context_lang = "Language: {}\n".format(cartridge_id)
-    context_info = prompter.perform_rag(query, os.path.join(cartridge_dir_path, "faiss.index"),
-                                        os.path.join(cartridge_dir_path, "chunks.txt"))
+
+    context_info: str = ""
+    document_chunks_path = os.path.join(cartridge_dir_path, "chunks.txt")
+    if chat_use_CAG:
+        context_info = prompter.perform_cag(query_joined, document_chunks_path)
+    else:
+        faiss_path = os.path.join(cartridge_dir_path, "faiss.index")
+        context_info = prompter.perform_rag(query_joined, faiss_path, document_chunks_path)
 
     context = context_lang + context_info
 
-    print("Context:\n" + context)
+    print("Query: {}\nContext:\n".format(query_joined) + context)
+
+    return context
+
+
+def cag_command(*query):
+    """
+    Experimental switch that toggles the use of 'Cache-augmented Generation'
+    over 'Retrieval-Augmented Generation'.
+    """
+    global chat_use_CAG
+
+    chat_use_CAG = not chat_use_CAG
+    print("TOGGLE: 'use_CAG' -> {}.".format(chat_use_CAG))
+
+
+def generate_command(*query):
+    """
+    Calls the chat model to generate based on context generated from
+    the query, and with a prompt specified in advance.
+    """
+    global last_cartridge_id, chat_prompt
+
+    if not prompter.check_model_server():
+        print("ERROR: Can't connect to model's server.")
+        return
+
+    print("===\n\n")
+    context = context_command(*query)
+    result = prompter.generate_response(chat_prompt, context)
+
+    print("Generated:\n" + result)
+    print("\n\n===")
 
 
 def exit_command():
+    """
+    Called before the program exits.
+    Could be used for deinitializing/cleanup.
+    """
     pass
 
 
 def help_command():
+    """
+    lists the available commands.
+    """
     global available_commands
 
     for c in available_commands.keys():
@@ -189,6 +262,7 @@ available_commands = {
     "unload": unload_command,
     "context": context_command,
     "generate": generate_command,
+    "cag": cag_command,
     "exit": exit_command
 }
